worker/requirements.txt
websocket-client>=1.7.0
azure-storage-blob>=12.19.0
requests>=2.31.0
